Replace int indexes, with long reference for put/get from the Memory interface.  Added FastUtil for use in unit tests

+++ src/test/java/com/lmax/disruptor/TypeValidator.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public interface TypeValidator
 {
     void init(Random r, int numberOfPuts, int size, int chunkSize);
-    boolean putAndGetAt(Memory memory, int i, int index, int offset);
-    boolean validateGetAt(Memory memory, int index, int offset);
+    boolean putAndGetAt(Memory memory, int i, long index, int offset);
+    boolean validateGetAt(Memory memory, long index, int offset);
     int getTypeSize();
 }
